--- src/main/java/subway/StationResponse.java
@@ -1,19 +1,0 @@
-package subway;
-
-public class StationResponse {
-    private Long id;
-    private String name;
-
-    public StationResponse(Long id, String name) {
-        this.id = id;
-        this.name = name;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
-    }
-}
